Moved tests to separate files, added new tests
